requirements.txt
django
django-ckeditor
django-jazzmin
mysqlclient
Pillow
django-admin-charts